Add files via upload
To get the program to work, make sure to compile the files on your computer using the 
line javac ArtHeist.java.

(Premise and how to play)
In this game you are a thief who is trying to steal from a modern art museum, but there
is a catch. That being the fact that all of the art is really strange and it is
therefore difficult to judge its value. You are given a prompt that asks you where you
want to go (to go to the location type the name of the location with exact casing) or if you want to steal something or check your inventory.
If you chose to steal an object, you get the item and the money that it is worth,
but you also increase your chances of getting caught by a guard. if you do get caught,
you can try to run away or throw an artwork at the guard. Throwing an item guarantees
an escape, but you will lose an artwork. Running will cost you noting but it is risky.
if you fail to run away twice, you will lose. To leave the museum, you need to make your
way back to the entrance and answer yes to the prompt you are given. You will then be
given an ending based on how much money you obtained.